pyproject.toml
[tool.poetry]
name = "dokku-api"
version = "0.1.0"
description = ""
authors = ["JeanExtreme002 <jeangamerextreme@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "0.19.2"
fastapi = "0.75.0"
pydantic = "1.10"
paramiko = "2.10.1"
python-decouple = "3.6"
anyio = "3.5.0"
asgiref = "3.5.0"
bcrypt = "3.2.0"
cffi = "1.15.0"
click = "8.0.4"
colorama = "0.4.4"
cryptography = "36.0.1"
h11 = "0.13.0"
httptools = "0.6.4"
idna = "3.3"
pycparser = "2.21"
PyNaCl = "1.5.0"
python-multipart = "0.0.5"
PyYAML = "6.0"
six = "1.16.0"
sniffio = "1.2.0"
starlette = "0.17.1"
typing_extensions = "4.1.1"
uvicorn = "0.17.5"
watchgod = "0.7"
websockets = "10.2"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^6.0.1"
coverage = "^7.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
